Move the app component to its own dir
