+++ src/app/App.js
@@ -1,5 +1,5 @@
 import "./App.css";
-import Success from "./pages/Sucess/Success.js";
+import Success from "../pages/Sucess/Success.js";
 import React from "react";
 
 function App() {
