+++ src/app/store.js
@@ -1,0 +1,8 @@
+import counterReducer from "../features/counter/counterSlice";
+import { configureStore } from "@reduxjs/toolkit";
+
+export const store = configureStore({
+  reducer: {
+    counter: counterReducer,
+  },
+});
